+++ components/style/values/computed/background.rs
@@ -5,12 +5,9 @@
 //! Computed types for CSS values related to backgrounds.
 
 use crate::values::computed::length::NonNegativeLengthOrPercentageOrAuto;
-use crate::values::computed::{Context, ToComputedValue};
 use crate::values::generics::background::BackgroundSize as GenericBackgroundSize;
-use crate::values::specified::background::BackgroundRepeat as SpecifiedBackgroundRepeat;
-use crate::values::specified::background::BackgroundRepeatKeyword;
-use std::fmt::{self, Write};
-use style_traits::{CssWriter, ToCss};
+
+pub use crate::values::specified::background::BackgroundRepeat;
 
 /// A computed value for the `background-size` property.
 pub type BackgroundSize = GenericBackgroundSize<NonNegativeLengthOrPercentageOrAuto>;
@@ -21,84 +18,6 @@ impl BackgroundSize {
         GenericBackgroundSize::Explicit {
             width: NonNegativeLengthOrPercentageOrAuto::auto(),
             height: NonNegativeLengthOrPercentageOrAuto::auto(),
-        }
-    }
-}
-
-/// The computed value of the `background-repeat` property:
-///
-/// https://drafts.csswg.org/css-backgrounds/#the-background-repeat
-#[derive(Clone, Debug, MallocSizeOf, PartialEq)]
-pub struct BackgroundRepeat(pub BackgroundRepeatKeyword, pub BackgroundRepeatKeyword);
-
-impl BackgroundRepeat {
-    /// Returns the `repeat repeat` value.
-    pub fn repeat() -> Self {
-        BackgroundRepeat(
-            BackgroundRepeatKeyword::Repeat,
-            BackgroundRepeatKeyword::Repeat,
-        )
-    }
-}
-
-impl ToCss for BackgroundRepeat {
-    fn to_css<W>(&self, dest: &mut CssWriter<W>) -> fmt::Result
-    where
-        W: Write,
-    {
-        match (self.0, self.1) {
-            (BackgroundRepeatKeyword::Repeat, BackgroundRepeatKeyword::NoRepeat) => {
-                dest.write_str("repeat-x")
-            },
-            (BackgroundRepeatKeyword::NoRepeat, BackgroundRepeatKeyword::Repeat) => {
-                dest.write_str("repeat-y")
-            },
-            (horizontal, vertical) => {
-                horizontal.to_css(dest)?;
-                if horizontal != vertical {
-                    dest.write_str(" ")?;
-                    vertical.to_css(dest)?;
-                }
-                Ok(())
-            },
-        }
-    }
-}
-
-impl ToComputedValue for SpecifiedBackgroundRepeat {
-    type ComputedValue = BackgroundRepeat;
-
-    #[inline]
-    fn to_computed_value(&self, _: &Context) -> Self::ComputedValue {
-        match *self {
-            SpecifiedBackgroundRepeat::RepeatX => BackgroundRepeat(
-                BackgroundRepeatKeyword::Repeat,
-                BackgroundRepeatKeyword::NoRepeat,
-            ),
-            SpecifiedBackgroundRepeat::RepeatY => BackgroundRepeat(
-                BackgroundRepeatKeyword::NoRepeat,
-                BackgroundRepeatKeyword::Repeat,
-            ),
-            SpecifiedBackgroundRepeat::Keywords(horizontal, vertical) => {
-                BackgroundRepeat(horizontal, vertical.unwrap_or(horizontal))
-            },
-        }
-    }
-
-    #[inline]
-    fn from_computed_value(computed: &Self::ComputedValue) -> Self {
-        // FIXME(emilio): Why can't this just be:
-        //   SpecifiedBackgroundRepeat::Keywords(computed.0, computed.1)
-        match (computed.0, computed.1) {
-            (BackgroundRepeatKeyword::Repeat, BackgroundRepeatKeyword::NoRepeat) => {
-                SpecifiedBackgroundRepeat::RepeatX
-            },
-            (BackgroundRepeatKeyword::NoRepeat, BackgroundRepeatKeyword::Repeat) => {
-                SpecifiedBackgroundRepeat::RepeatY
-            },
-            (horizontal, vertical) => {
-                SpecifiedBackgroundRepeat::Keywords(horizontal, Some(vertical))
-            },
         }
     }
 }
